+++ src/main/fc/fc_core.h
@@ -46,5 +46,8 @@ bool processRx(timeUs_t currentTimeUs);
 void updateArmingStatus(void);
 void updateRcCommands(void);
 
-void taskMainPidLoop(timeUs_t currentTimeUs);
 bool isFlipOverAfterCrashMode(void);
+
+#if defined(USE_GPS) || defined(USE_MAG)
+void updateMagHold(void);
+#endif
